Join a given string a number of times
The function will join the user provided string n number of times where
n is the number provided by the user too.
THe function ensures that it works even if blank string is provided.
The function throws an error message if a mischievous user provides
negative value to the number of times strings need to be multiplied.