refactor(Checking if validity of passed arguements): Refactor the checking of the arguements
Use array.every method to check every element for a specific condition,
in this case, check if these are postive and are indeed numbers;The loop
to sum the all the numbers in the range runs only after this initial
checking. its wortjwhile to note that foreach method wont work, instead
every method is needed to check each elment in a list for specific
conditions.