feat(Function to check if given year is a leap year): Check if a year is leap year
First check if divisble by 4. If not divisble by 4, then straight away
declare not a leap year. But if divisible by 4 then first check if also
divisible ny 400 (then it is indeed a leap year), however, if 4
divisible year is not divsible by 400, it can be still bea leap year
only if it is also not divisble by 100.